Map M-h to personal prefix map
So now there're 2 ways to access personal prefix: C-h or M-h. Outside of
god-mode, it's easier to use M-h. This is the reason for the binding. This
replaces `mark-paragraph` so I rebind `mark-paragraph` to "M-h h". In god-mode,
it's only "h h", so still very convenient. When I want to call it repeatedly, I
can just spam `.`.